+++ libkernelflinger/qsort.c
@@ -144,7 +144,7 @@ loop:	SWAPINIT(a, es);
 	pn = (char *)a + n * es;
 	r = min(pa - (char *)a, pb - pa);
 	vecswap(a, pb - r, r);
-	r = min((size_t)(pd - pc), pn - pd - es);
+	r = min(((size_t)pd - (size_t)pc), pn - pd - es);
 	vecswap(pb, pn - r, r);
 	if ((r = pb - pa) > es)
 		qsort(a, r / es, es, cmp);
